Do not push to the answering device of a call
This avoids an error where the answering device hangs up immediately
because it also receives the push that the CommunicationRequest was
set to completed

When a call is answered, the recipient that answers the call sets the status
of the CommunicationRequest to 'completed'. That triggers a push to the other
recipients so that they can stop ringing. This push was also sent to the recipient
that answered the call and therefore made the client to immediately hang up the
call.

Push workflow can be found here:
https://cloud.awesome-technologies.de/index.php/apps/drawio/511281

Change-Id: I4c33dc52b1e1e46af93725bf8d655a826be68979