Test authorization for org to read different resources
Creates a bunch of patients and either ServiceRequests or
Communications between the organizations according to a simple
pattern and checks that access to patients is either denied
or allowed based on the existence of matching Communication
or ServiceRequest resources. Access from org to the
individual Communication and ServiceRequests are also
checked.

Change-Id: I10a499bf6a263d3913a6be4d60fba25d4adb79c0