Add perf measurement interceptor
It takes measurements before and after the other code that handles
a few of the available pointcuts. At the end of a request, it will
output to the logger as an "info" message, including the time from
start of request to the start of a given pointcut, and the time
from the very start to the very end.

It starts at the "incoming request post processed" pointcut, as
at the "request pre processed" point, there is almost nothing
available from the request itself, as neither the url nor the
http verb have been parsed yet.

Change-Id: I08ff4d251072fb000a25fb16493b4cb474dac157